% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_data_irs_web}
\alias{check_data_irs_web}
\title{Check data availability for IRS and website}
\usage{
check_data_irs_web(ein, year = 2019, sources = c("irs", "website"))
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{year}{A year in which a form was filed. The default value is 2019.}

\item{sources}{The data sources. The default value is \code{c("irs", "website")}.}
}
\value{
A datafram that contains three columns: EIN, its data source and availability. Data availability column is a dummy variable. 1 = data exist. 0 = data don't exist.
}
\description{
Check data availability for IRS and website
}
