% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_populate_grantmaking.r
\name{check_for_grantmaking_activity_990}
\alias{check_for_grantmaking_activity_990}
\title{Check for grant-making activity from 990 forms}
\usage{
check_for_grantmaking_activity_990(xml_root)
}
\arguments{
\item{xml_root}{An XML root element associated with a particular organization}
}
\value{
The function either returns the information on grant-making activity (numeric) or states that "This organization did not file ScheduleI."
}
\description{
Check for grant-making activity from 990 forms
}
