% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texts.R
\name{clean_program_desc}
\alias{clean_program_desc}
\title{Clean the descriptions of the programs run by a particular organization}
\usage{
clean_program_desc(program_desc, text_length_threshold)
}
\arguments{
\item{program_desc}{Program descriptions}

\item{text_length_threshold}{The length of the minimum words to filter the program descriptions}
}
\value{
If successful, the function returns the cleaned descriptions of the programs run by a particular organization
}
\description{
Clean the descriptions of the programs run by a particular organization
}
