% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{combine_texts}
\alias{combine_texts}
\title{Combine multiple webpage texts into one shared text with headers and footers removed}
\usage{
combine_texts(x)
}
\arguments{
\item{x}{A vector of texts}
}
\value{
combined_text A data frame with three columns. Header is the shared header
text. Footer is the shared footer text. Body is the concatenated text with header and
footer removed
}
\description{
Combine multiple webpage texts into one shared text with headers and footers removed
}
