% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{extract_about_page_content}
\alias{extract_about_page_content}
\title{Extract about page content}
\usage{
extract_about_page_content(about_url)
}
\arguments{
\item{about_url}{An URL for an about page}
}
\value{
About page text (character vector)
}
\description{
Extract about page content
}
