% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_contacts.r
\name{extract_contact_page_content}
\alias{extract_contact_page_content}
\title{Extract contact page content}
\usage{
extract_contact_page_content(contact_url)
}
\arguments{
\item{contact_url}{An URL for an contact page}
}
\value{
contact page text (character vector)
}
\description{
Extract contact page content
}
