% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_populate_grantmaking.r
\name{extract_financial_info}
\alias{extract_financial_info}
\title{Extract financial information from 990 forms}
\usage{
extract_financial_info(xml_plucked, variable)
}
\arguments{
\item{xml_plucked}{A content element of the 990 XML file}

\item{variable}{A particular field in 990 forms}
}
\value{
If successful, the function returns the selected financial information from a 990 form in XML format (numeric data type).
}
\description{
Extract financial information from 990 forms
}
