% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_populate_grantmaking.r
\name{filter_null_grant_info}
\alias{filter_null_grant_info}
\title{Filter null grant information}
\usage{
filter_null_grant_info(variable)
}
\arguments{
\item{variable}{A particular field in 990 forms}
}
\value{
The function either returns the numeric financial information or 0 (in the case of NULL).
}
\description{
Filter null grant information
}
