% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{find_about_link}
\alias{find_about_link}
\title{Find links and other information related to about page}
\usage{
find_about_link(base_url)
}
\arguments{
\item{base_url}{A base URL (the base part of the web address)}
}
\value{
If successful, the function returns a working base URL. If not, the function reports one of the following four error cases: "This link does not have about page.", "This link has a PHP error.", "This link is flat (not tree structure).", or "This link is broken."
}
\description{
Find links and other information related to about page
}
