% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter_handles.r
\name{find_all}
\alias{find_all}
\title{Find Twitter handles either using a name (default strategy) or an website of an organization.}
\usage{
find_all(org_name, website_address = NULL)
}
\arguments{
\item{org_name}{A name of an organization}

\item{website_address}{An website of an organization. The default value is NULL.}
}
\value{
If no website information were provided, the function returns the Twitter handles discovered via Bing API. If website information were provided, the function returns a dataframe that contains Twitter handles discovered by Bing API ("Bing" column), the website ("Website" column), and the intersection of the two search results ("Intersection" column).
}
\description{
Find Twitter handles either using a name (default strategy) or an website of an organization.
}
