% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facebook_handles.r
\name{find_facebook_page_from_org_page}
\alias{find_facebook_page_from_org_page}
\title{Parse Facebook handles from an organization website}
\usage{
find_facebook_page_from_org_page(website_address)
}
\arguments{
\item{website_address}{An website address of an organization}
}
\value{
A list of Facebook handles
}
\description{
Parse Facebook handles from an organization website
}
