% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{find_footer}
\alias{find_footer}
\title{Find a common footer string in a set of texts}
\usage{
find_footer(x, threshold = 0.9)
}
\arguments{
\item{x}{A vector of texts}

\item{threshold}{Fraction of the texts required to share the common footer string}
}
\value{
footer A string that is the longest common footer string for all the texts.
}
\description{
Find a common footer string in a set of texts
}
