% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{find_header}
\alias{find_header}
\title{Find a common header string in a set of texts}
\usage{
find_header(x, threshold = 0.9)
}
\arguments{
\item{x}{A vector of texts}

\item{threshold}{Fraction of the texts required to share the common header string}
}
\value{
header A string that is the longest common header string for all the texts.
}
\description{
Find a common header string in a set of texts
}
