% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_page_info.R
\name{find_possible_link}
\alias{find_possible_link}
\title{Find a possible link based on a base URL}
\usage{
find_possible_link(base_url, possible_link)
}
\arguments{
\item{base_url}{base URL (the base part of the web address)}

\item{href}{possible_link Possible link}
}
\value{
If successful, the function returns a dataframe of three columns ("href", "link_text", "href"). href is "base." link_text is "Found without tree search." href is the URL. If not successful, the result is NA.
}
\description{
Find a possible link based on a base URL
}
