% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_page_info.R
\name{find_search_term}
\alias{find_search_term}
\title{Check whether the href attribute contains "search_term" or not}
\usage{
find_search_term(about_links, search_term, base_url = "http://example.com")
}
\arguments{
\item{about_links}{A dataframe of three columns ("href", "link_text", and "link").}

\item{search_term}{A search term.}

\item{base_url}{A base URL (the base part of the web address)}
}
\value{
A dataframe of three columns ("href", "link_text", and "link") filtered by a search term and cleaned for later use in the workflow.
}
\description{
Check whether the href attribute contains "search_term" or not
}
