% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter_handles.r
\name{find_twitter_handle_from_bing}
\alias{find_twitter_handle_from_bing}
\title{Find Twitter handles using Bing API}
\usage{
find_twitter_handle_from_bing(org_name, bing_search_api_key)
}
\arguments{
\item{org_name}{An organization name}

\item{bing_search_api_key}{A Bing search API key (Note that the base URL is "https://api.cognitive.microsoft.com/bing/v7.0/search")}
}
\value{
A list of Twitter handles
}
\description{
Find Twitter handles using Bing API
}
