% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter_handles.r
\name{find_twitter_handle_from_org_page}
\alias{find_twitter_handle_from_org_page}
\title{Parse Twitter handles from an organization website}
\usage{
find_twitter_handle_from_org_page(website_address)
}
\arguments{
\item{website_address}{An website address of an organization}
}
\value{
A list of Twitter handles
}
\description{
Parse Twitter handles from an organization website
}
