% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R
\name{get_990}
\alias{get_990}
\title{Get the XML root element associated with a particular Employment Identification Numbers}
\usage{
get_990(ein)
}
\arguments{
\item{ein}{An Employment Identification Numbers}
}
\value{
If successful, the function returns the XML root element associated with a particular Employment Identification Numbers
}
\description{
Get the XML root element associated with a particular Employment Identification Numbers
}
