% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{get_about_page_content}
\alias{get_about_page_content}
\title{Get about page content}
\usage{
get_about_page_content(base_url)
}
\arguments{
\item{base_url}{A base URL (the base part of the web address)}
}
\value{
about_page A dataframe that contains about page text
}
\description{
Get about page content
}
