% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texts.R
\name{get_all_financial_data}
\alias{get_all_financial_data}
\title{Get all financial fields for a given org}
\usage{
get_all_financial_data(xml_root)
}
\arguments{
\item{xml_root}{An XML root element associated with a particular organization}
}
\value{
A data frame with 30 financial fields
}
\description{
Get all financial fields for a given org
}
