% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{get_all_texts}
\alias{get_all_texts}
\title{Extract all about page texts for a given organization}
\usage{
get_all_texts(org_url)
}
\arguments{
\item{org_url}{A URL (the web address of the organization of interest}
}
\value{
about_texts A dataframe that contains the combined extracted text from the
homepage and about pages, the shared header of the pages if found, and the shared
footer if found.
}
\description{
Extract all about page texts for a given organization
}
