% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_contacts.r
\name{get_contact_links_from_website}
\alias{get_contact_links_from_website}
\title{Get child links associated with email contacts}
\usage{
get_contact_links_from_website(base_url)
}
\arguments{
\item{base_url}{A website URL}
}
\value{
out A character vector that contains child links associated with email contacts (i.e., page_urls)
}
\description{
Get child links associated with email contacts
}
