% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_contacts.r
\name{get_contact_page_content}
\alias{get_contact_page_content}
\title{Get contact page content}
\usage{
get_contact_page_content(base_url)
}
\arguments{
\item{base_url}{A base URL (the base part of the web address)}
}
\value{
contact_page A dataframe that contains email contact
}
\description{
Get contact page content
}
