% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_contacts.r
\name{get_email_contact_from_webpage}
\alias{get_email_contact_from_webpage}
\title{Get contact webpage content}
\usage{
get_email_contact_from_webpage(page_url)
}
\arguments{
\item{page_url}{A webpage URL}
}
\value{
out A dataframe that contains email contact
}
\description{
Get contact webpage content
}
