% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_contacts.r
\name{get_emails_from_website}
\alias{get_emails_from_website}
\title{Extract emails from a base url}
\usage{
get_emails_from_website(base_url)
}
\arguments{
\item{base_url}{A website URL}
}
\value{
out A dataframe that contains the page URLs and the email addresses appearing on these URLs
}
\description{
Extract emails from a base url
}
