% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R,
%   R/org_classify_description_texts.R
\name{get_filing_type_990}
\alias{get_filing_type_990}
\title{Get 990 filing type}
\usage{
get_filing_type_990(xml_root)

get_filing_type_990(xml_root)
}
\arguments{
\item{xml_root}{An XML root element associated with a particular organization}
}
\description{
Get 990 filing type

Get 990 filing type
}
