% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R,
%   R/org_classify_description_texts.R
\name{get_foundation_ein}
\alias{get_foundation_ein}
\title{Get Employer Identification Numbers (EINs) associated with foundations}
\usage{
get_foundation_ein()

get_foundation_ein()
}
\arguments{
\item{idx_year}{A XML file it contains the 990 forms filed in a particular year. An outcome of import_idx() function.}
}
\value{
If successful, the function returns the EINs associated with foundations.

If successful, the function returns the EINs associated with foundations.
}
\description{
Get Employer Identification Numbers (EINs) associated with foundations

Get Employer Identification Numbers (EINs) associated with foundations
}
