% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_populate_grantmaking.r
\name{get_grantmaking_details_990}
\alias{get_grantmaking_details_990}
\title{Get grant-making details from 990 forms}
\usage{
get_grantmaking_details_990(xml_root)
}
\arguments{
\item{xml_root}{An XML root element associated with a particular organization}
}
\value{
If successful, the function returns a dataframe of six columns that contains information the amount of various grants the organization received. If such information were absent in the data, then these columns would contain only NA values.
}
\description{
Get grant-making details from 990 forms
}
