% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_page_info.R
\name{get_links}
\alias{get_links}
\title{Get links base on base URL}
\usage{
get_links(base_url)
}
\arguments{
\item{base_url}{A base URL (the base part of the web address)}
}
\value{
The function returns a dataframe of three columns ("href", "link_text", and "link"). If the function runs into an error getting the links based on the base URL, the value of the href column should be NA. The link_text will inform the error case. This should be either "This website is broken.", "The website is flat (no tree structure)." or "The website does not have about page."
}
\description{
Get links base on base URL
}
