% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texts.R
\name{get_organization_ein}
\alias{get_organization_ein}
\title{Search for the EIN an organization given a name and U.S. state}
\usage{
get_organization_ein(state_abbrev, org_name)
}
\arguments{
\item{state_abbrev}{A two-letter U.S. state abbreviation (e.g., CA)}

\item{org_name}{The name of a nonprofit organization}
}
\value{
If successful, the function returns the Employment Identification Number matching the organization name in the given state
}
\description{
Search for the EIN an organization given a name and U.S. state
}
