% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texts.R
\name{get_organization_name_990}
\alias{get_organization_name_990}
\title{Get the name of the organization associated with a particular Employment Identification Numbers}
\usage{
get_organization_name_990(ein)
}
\arguments{
\item{ein}{An Employment Identification Numbers}
}
\value{
If successful, the function returns the name of the organization associated with a particular Employment Identification Numbers
}
\description{
Get the name of the organization associated with a particular Employment Identification Numbers
}
