% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R
\name{get_scheduleO}
\alias{get_scheduleO}
\title{Get concrete information from Schedule O}
\usage{
get_scheduleO(ein)
}
\arguments{
\item{ein}{An Employment Identification Numbers}
}
\value{
The function returns either concrete information from Schedule O (likely a character vector) or states that such information is not present.
}
\description{
Get concrete information from Schedule O
}
