% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texts.R
\name{get_single_value_990}
\alias{get_single_value_990}
\title{Get concrete information from 990 forms}
\usage{
get_single_value_990(xml_root, irs_variable)
}
\arguments{
\item{xml_root}{An XML root element associated with a particular organization}

\item{irs_variable}{A type of concrete information. It should correspond to a package_variable in the package's irs_fields table}
}
\value{
Depending on the type parameter, the function returns the value of the 990 form for the given variable or else NA.
}
\description{
Get concrete information from 990 forms
}
