% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{if_not_about}
\alias{if_not_about}
\title{Check whether the href attribute contains "about" or not}
\usage{
if_not_about(href)
}
\arguments{
\item{href}{An href attribute}
}
\value{
Either an href attribute (YES) or NA (NO)
}
\description{
Check whether the href attribute contains "about" or not
}
