% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_contacts.r
\name{if_not_contact}
\alias{if_not_contact}
\title{Check whether the href attribute contains "contact" or not}
\usage{
if_not_contact(href)
}
\arguments{
\item{href}{An href attribute}
}
\value{
Either an href attribute (YES) or NA (NO)
}
\description{
Check whether the href attribute contains "contact" or not
}
