% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_page_info.R
\name{if_not_search_term}
\alias{if_not_search_term}
\title{Check whether the href attribute contains a user-defined "search_term"}
\usage{
if_not_search_term(href, search_term)
}
\arguments{
\item{href}{An href attribute}
}
\value{
Either an href attribute 1 (\code{YES}) or NA (\code{NA})
}
\description{
Check whether the href attribute contains a user-defined "search_term"
}
