% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R,
%   R/org_classify_description_texts.R
\name{ifnotNA}
\alias{ifnotNA}
\title{Simplify ifelse process}
\usage{
ifnotNA(var)

ifnotNA(var)
}
\arguments{
\item{var}{target variable (likely character type)}
}
\value{
if the var length is 0 then the function returns NA, otherwise it returns the input parameter.

if the var length is 0 then the function returns NA, otherwise it returns the input parameter.
}
\description{
Simplify ifelse process

Simplify ifelse process
}
