% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R
\name{import_idx}
\alias{import_idx}
\title{Import machine-readable data from 990 forms filed with the IRS}
\usage{
import_idx(year)
}
\arguments{
\item{year}{A year in which a form was filed.}
}
\value{
If successful, the function returns a XML file it contains the 990 forms filed in a particular year.
}
\description{
Import machine-readable data from 990 forms filed with the IRS
}
