% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R,
%   R/org_classify_description_texts.R
\docType{data}
\name{irs_fields}
\alias{irs_fields}
\title{Data fields available from IRS 990 forms}
\format{
A data frame with 32 rows and 9 variables:
\describe{
\item{package_variable}{name to use in this package to retrieve the field}
\item{category}{a top-level variable the variable belongs to}
\item{subcategory}{a subcategory the variable belongs to}
\item{XML_990}{the XML location for the parsed 990 form}
\item{Form_990}{the Part and Line number of the form location for this XML}
\item{XML_990EZ}{the XML location for the parsed 990 form}
\item{Form_990EZ}{the Part and Line number of the form location for this XML}
\item{XML_990PF}{the XML location for the parsed 990 form}
\item{Form_990PF}{the Part and Line number of the form location for this XML}
}

A data frame with 32 rows and 9 variables:
\describe{
\item{package_variable}{name to use in this package to retrieve the field}
\item{category}{a top-level variable the variable belongs to}
\item{subcategory}{a subcategory the variable belongs to}
\item{XML_990}{the XML location for the parsed 990 form}
\item{Form_990}{the Part and Line number of the form location for this XML}
\item{XML_990EZ}{the XML location for the parsed 990 form}
\item{Form_990EZ}{the Part and Line number of the form location for this XML}
\item{XML_990PF}{the XML location for the parsed 990 form}
\item{Form_990PF}{the Part and Line number of the form location for this XML}
}
}
\usage{
irs_fields

irs_fields
}
\description{
A dataset containing the variable names, associated Form 990 XML locations, Form 990 actual locations, and info
for retrievable 990 Fields

A dataset containing the variable names, associated Form 990 XML locations, Form 990 actual locations, and info
for retrievable 990 Fields
}
\keyword{datasets}
