% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_about_pages.r
\name{parse_by_length}
\alias{parse_by_length}
\title{Extract significant text from html source}
\usage{
parse_by_length(url, sentence_threshold = 5)
}
\arguments{
\item{url}{A URL (the web address to extract text from)}

\item{sentence_threshold}{an integer signifying the minimum word length for an included sentence}
}
\value{
extracted_text The text found on a page
}
\description{
Extract significant text from html source
}
