% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facebook_handles.r
\name{parse_facebook_page_from_page}
\alias{parse_facebook_page_from_page}
\title{Parse Facebook handles from a webpage}
\usage{
parse_facebook_page_from_page(website_address, google_search_results = NULL)
}
\arguments{
\item{website_address}{An website address of an organization}

\item{google_search_results}{The default is NULL. If you use a Google search result, use a non-NULL value.}
}
\value{
A list of facebook pages
}
\description{
Parse Facebook handles from a webpage
}
