% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter_handles.r
\name{parse_twitter_handle_from_page}
\alias{parse_twitter_handle_from_page}
\title{Parse Twitter handles from a webpage}
\usage{
parse_twitter_handle_from_page(website_address, google_search_results = NULL)
}
\arguments{
\item{website_address}{An website address of an organization}

\item{google_search_results}{The default is NULL. If you use a Google search result, use a non-NULL value.}
}
\value{
A list of Twitter handles
}
\description{
Parse Twitter handles from a webpage
}
