% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_page_info.R
\name{search_all_activities}
\alias{search_all_activities}
\title{Search all activities}
\usage{
search_all_activities(base_url)
}
\arguments{
\item{base_url}{A base URL (the base part of the web address)}
}
\value{
A data.frame that contains the binary outcomes of the 13 search results: donations, events, membership, digital membership, chapters, volunteers, services, taking actions, advocacy, visits, resources, boards, and press.
}
\description{
Search all activities
}
