% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_populate_grantmaking.r
\name{standardize_990_flag}
\alias{standardize_990_flag}
\title{Standardize 990 flag}
\usage{
standardize_990_flag(flag_value)
}
\arguments{
\item{flag_value}{A value associated with 990 flag}
}
\value{
The function makes the flag_value either 1 (flag_value == "true") or 0 (other cases plus when length(flag_value) == 0).
}
\description{
Standardize 990 flag
}
