% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texts.R
\name{standardize_url}
\alias{standardize_url}
\title{Standardize the website URL of an organization}
\usage{
standardize_url(raw_website)
}
\arguments{
\item{raw_website}{The raw website URL of an organization}
}
\value{
If successful, the function returns the standardized version of the raw website URL.
}
\description{
Standardize the website URL of an organization
}
