% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_data_availability}
\alias{check_data_availability}
\title{Check data availability}
\usage{
check_data_availability(
  ein,
  year = 2019,
  source = c("irs", "website", "social_media")
)
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{year}{A year in which a form was filed. The default value is 2019.}

\item{source}{The data source. Four options exist: "irs", "website", "twitter", "facebook."}
}
\value{
A datafram that contains three columns: EIN, its data source and availability. Data availability column is a dummy variable. 1 = data exist. 0 = data don't exist.
}
\description{
Check data availability
}
