% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R,
%   R/org_classify_description_texts.R
\name{get_990}
\alias{get_990}
\title{Get the XML root element associated with a particular Employment Identification Numbers}
\usage{
get_990(ein, year = 2019)

get_990(ein, year = 2019)
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{year}{A year in which a form was filed. The default value is 2019.}
}
\value{
If successful, the function returns the XML root element associated with a particular Employment Identification Numbers

If successful, the function returns the XML root element associated with a particular Employment Identification Numbers
}
\description{
Get the XML root element associated with a particular Employment Identification Numbers

Get the XML root element associated with a particular Employment Identification Numbers
}
