% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R,
%   R/org_classify_description_texts.R
\name{get_aws_url}
\alias{get_aws_url}
\title{Get the Amazon Web Server URL associated with a particular Employment Identification Numbers}
\usage{
get_aws_url(ein, year = 2019)

get_aws_url(ein, year = 2019)
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{year}{A year in which a form was filed. The default IRS year is 2019.}

\item{form}{An IRS document form. The default is NULL. There are three other options: "990", "990PF", "990EZ"}

\item{tax_period}{A tax period (year-month). The default tax period is 2019.}

\item{move_global}{Whether moving the XML file, which contains the 990 forms filed in a particular year, to the user's global environment. The default value is TRUE.}

\item{multiple_tax_period}{This is only applied to the case when an EIN is associated with multiple tax periods. If you want to extract the latest tax period's latest submission, set this argument to FALSE. If you want to extract the latest submission from the multiple tax periods, set this argument to TRUE.}
}
\value{
If a single tax period was identified, the function returns the Amazon Web Server URL associated with a particular Employment Identification Numbers. If multiple tax periods were identified, the function returns either a vector that only includes the URL (multiple_tax_period= FALSE) or a dataframe that includes the URL as well as related tax periods (multiple_tax_period = TRUE).

If successful, the function returns the Amazon Web Server URL associated with a particular Employment Identification Numbers.
}
\description{
Get the Amazon Web Server URL associated with a particular Employment Identification Numbers

Get the Amazon Web Server URL associated with a particular Employment Identification Numbers
}
