% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R
\name{get_aws_url}
\alias{get_aws_url}
\title{Get the Amazon Web Server URL associated with a particular Employment Identification Numbers}
\usage{
get_aws_url(ein, form = NULL)
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{form}{The default is NULL. There are three other options: "990", "990PF", "990EZ"}
}
\value{
If successful, the function returns the Amazon Web Server URL associated with a particular Employment Identification Numbers.
}
\description{
Get the Amazon Web Server URL associated with a particular Employment Identification Numbers
}
