% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texts.R
\name{get_aws_url}
\alias{get_aws_url}
\title{Get the Amazon Web Server URL associated with a particular Employment Identification Numbers}
\usage{
get_aws_url(ein, year = 2019)
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{year}{A year in which a form was filed. The default IRS year is 2019.}
}
\value{
If successful, the function returns the Amazon Web Server URL associated with a particular Employment Identification Numbers.
}
\description{
Get the Amazon Web Server URL associated with a particular Employment Identification Numbers
}
