% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R,
%   R/org_classify_description_texts.R
\name{get_organization_name_990}
\alias{get_organization_name_990}
\title{Get the name of the organization associated with a particular Employment Identification Numbers}
\usage{
get_organization_name_990(ein)

get_organization_name_990(ein)
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{idx_year}{A XML file it contains the 990 forms filed in a particular year. An outcome of import_idx() function.}
}
\value{
If successful, the function returns the name of the organization associated with a particular Employment Identification Numbers

If successful, the function returns the name of the organization associated with a particular Employment Identification Numbers
}
\description{
Get the name of the organization associated with a particular Employment Identification Numbers

Get the name of the organization associated with a particular Employment Identification Numbers
}
