% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_classify_description_texs.R
\name{get_scheduleR}
\alias{get_scheduleR}
\title{Get concrete information from Schedule R}
\usage{
get_scheduleR(ein, year = 2019)
}
\arguments{
\item{ein}{An Employment Identification Numbers}

\item{year}{A year in which a form was filed. The default value is 2019.}
}
\value{
The function returns either concrete information from Schedule R (likely a character vector) or states that such information is not present.
}
\description{
Get concrete information from Schedule R
}
